AdvzInternal struct members type usize -> u32 (yikes)

+++ primitives/src/vid/advz/payload_prover.rs
@@ -97,7 +97,7 @@ where
         let elems_iter = bytes_to_field::<_, KzgEval<E>>(&payload[range_poly_byte]);
         let mut proofs = Vec::with_capacity(range_poly.len() * points.len());
         for (i, evals_iter) in elems_iter
-            .chunks(self.recovery_threshold)
+            .chunks(self.recovery_threshold as usize)
             .into_iter()
             .enumerate()
         {
@@ -262,10 +262,11 @@ where
             .chain(proof.suffix_elems.iter().cloned());
 
         // rebuild the poly commits, check against `common`
-        for (commit_index, evals_iter) in range_poly
-            .into_iter()
-            .zip(elems_iter.chunks(self.recovery_threshold).into_iter())
-        {
+        for (commit_index, evals_iter) in range_poly.into_iter().zip(
+            elems_iter
+                .chunks(self.recovery_threshold as usize)
+                .into_iter(),
+        ) {
             let poly = self.polynomial(evals_iter);
             let poly_commit = UnivariateKzgPCS::commit(&self.ck, &poly).map_err(vid)?;
             if poly_commit != stmt.common.poly_commits[commit_index] {
@@ -294,18 +295,18 @@ where
         }
     }
     fn range_elem_to_poly(&self, range: &Range<usize>) -> Range<usize> {
-        range_coarsen(range, self.recovery_threshold)
+        range_coarsen(range, self.recovery_threshold as usize)
     }
     fn range_byte_to_poly(&self, range: &Range<usize>) -> Range<usize> {
         range_coarsen(
             range,
-            self.recovery_threshold * elem_byte_capacity::<KzgEval<E>>(),
+            self.recovery_threshold as usize * elem_byte_capacity::<KzgEval<E>>(),
         )
     }
     fn range_poly_to_byte_clamped(&self, range: &Range<usize>, len: usize) -> Range<usize> {
         let result = range_refine(
             range,
-            self.recovery_threshold * elem_byte_capacity::<KzgEval<E>>(),
+            self.recovery_threshold as usize * elem_byte_capacity::<KzgEval<E>>(),
         );
         Range {
             end: ark_std::cmp::min(result.end, len),
@@ -315,12 +316,12 @@ where
     fn offset_poly_to_elem(&self, range_poly_start: usize, range_elem_start: usize) -> usize {
         let start_poly_byte = index_refine(
             range_poly_start,
-            self.recovery_threshold * elem_byte_capacity::<KzgEval<E>>(),
+            self.recovery_threshold as usize * elem_byte_capacity::<KzgEval<E>>(),
         );
         range_elem_start - index_coarsen(start_poly_byte, elem_byte_capacity::<KzgEval<E>>())
     }
     fn final_poly_points_range_end(&self, range_elem_len: usize, offset_elem: usize) -> usize {
-        (range_elem_len + offset_elem - 1) % self.recovery_threshold + 1
+        (range_elem_len + offset_elem - 1) % self.recovery_threshold as usize + 1
     }
 
     fn check_stmt_consistency(stmt: &Statement<Self>) -> VidResult<()> {
@@ -406,9 +407,9 @@ mod tests {
         let num_random_cases = 20;
 
         // more items as a function of the above
-        let payload_elems_len = num_polys * recovery_threshold;
+        let payload_elems_len = num_polys * recovery_threshold as usize;
         let payload_bytes_base_len = payload_elems_len * elem_byte_capacity::<E::ScalarField>();
-        let poly_bytes_len = recovery_threshold * elem_byte_capacity::<E::ScalarField>();
+        let poly_bytes_len = recovery_threshold as usize * elem_byte_capacity::<E::ScalarField>();
         let mut rng = jf_utils::test_rng();
         let srs = init_srs(payload_elems_len, &mut rng);
         let mut advz =
